# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Turismo Ecuador API",
        "description": "API para el manejo de información turística de Ecuador",
        "contact": {
            "email": "soporte@turismo-ec.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8081",
            "description": "API Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/api/atractivos": {
            "get": {
                "tags": [
                    "Attractions"
                ],
                "summary": "Get all active attractions",
                "operationId": "c2f937cc5f6edbc5042fbe13dd223f24",
                "responses": {
                    "200": {
                        "description": "List of active attractions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Atractivo"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Attractions"
                ],
                "summary": "Create a new attraction",
                "operationId": "b6a90baf0c9763e9e9be71ffdcfc3a07",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "categoria_id",
                                    "provincia_id"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Volcán Cotopaxi"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Uno de los volcanes activos más altos del mundo"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Parque Nacional Cotopaxi"
                                    },
                                    "latitud": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "-0.6836"
                                    },
                                    "longitud": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "-78.4372"
                                    },
                                    "horario_visitas": {
                                        "type": "string",
                                        "example": "08:00 - 17:00"
                                    },
                                    "precio_entrada": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10.5
                                    },
                                    "estado": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "categoria_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "provincia_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Attraction created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Atractivo"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/atractivos/{id}": {
            "get": {
                "tags": [
                    "Attractions"
                ],
                "summary": "Get a specific attraction",
                "operationId": "217587dd8852815b1a0265730a64f62c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Attraction ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Attraction details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Atractivo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Attraction not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Attractions"
                ],
                "summary": "Update an attraction",
                "operationId": "521671a52b4a49e6ca329855fdc41f5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Attraction ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Volcán Cotopaxi Actualizado"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Nueva descripción actualizada"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Nueva dirección"
                                    },
                                    "latitud": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "-0.6836"
                                    },
                                    "longitud": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "-78.4372"
                                    },
                                    "horario_visitas": {
                                        "type": "string",
                                        "example": "09:00 - 18:00"
                                    },
                                    "precio_entrada": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 15
                                    },
                                    "estado": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "categoria_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "provincia_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Attraction updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Atractivo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Attraction not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Attractions"
                ],
                "summary": "Delete an attraction",
                "operationId": "1475c0f2083b32f1e16b16ab4ca26dd3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Attraction ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Attraction deleted successfully"
                    },
                    "404": {
                        "description": "Attraction not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/categorias": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get all categories",
                "operationId": "5b1f217a1898f32c344cecca1384320d",
                "responses": {
                    "200": {
                        "description": "List of categories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Categoria"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Create a new category",
                "operationId": "ec8a5f1a1e31435147a2b0530f963838",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Aventura"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Actividades de aventura y deportes extremos"
                                    },
                                    "estado": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Category created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Categoria"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/categorias/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get a specific category",
                "operationId": "06dae9190b206363656a24ddd8e80f5c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Categoria"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Update a category",
                "operationId": "ebdb99296b27bc0de05e0553d86ce801",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Aventura Actualizada"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Nueva descripción"
                                    },
                                    "estado": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Category updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Categoria"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Delete a category",
                "operationId": "65254db064fddaaa5a532d92552c8724",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category deleted successfully"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/contactos": {
            "get": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Get all contacts",
                "operationId": "1b18600c9907dce83c555a8b51463b99",
                "responses": {
                    "200": {
                        "description": "List of contacts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Contacto"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Create a new contact",
                "operationId": "ee65bc5bf7b742b7d785518bf53fcfda",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "email",
                                    "mensaje"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@example.com"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "+593 99 999 9999"
                                    },
                                    "mensaje": {
                                        "type": "string",
                                        "example": "Me gustaría obtener más información sobre..."
                                    },
                                    "asunto": {
                                        "type": "string",
                                        "example": "Consulta general"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contact created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contacto"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
        "/api/provincias": {
            "get": {
                "tags": [
                    "Provinces"
                ],
                "summary": "Get all provinces",
                "operationId": "09f2d238a699516225ec68b8f9596e30",
                "responses": {
                    "200": {
                        "description": "List of provinces",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Provincia"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Provinces"
                ],
                "summary": "Create a new province",
                "operationId": "cd75aa89e80e6b4297db26ad91b57f2a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "codigo_iso",
                                    "region"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Pichincha"
                                    },
                                    "codigo_iso": {
                                        "type": "string",
                                        "example": "P"
                                    },
                                    "region": {
                                        "type": "string",
                                        "example": "Sierra"
                                    },
                                    "poblacion": {
                                        "type": "integer",
                                        "example": 3000000
                                    },
                                    "superficie": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 9416
                                    },
                                    "capital": {
                                        "type": "string",
                                        "example": "Quito"
                                    },
                                    "idioma_principal": {
                                        "type": "string",
                                        "example": "Español"
                                    },
                                    "estado": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Province created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Provincia"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Atractivo": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Volcán Cotopaxi"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Uno de los volcanes activos más altos del mundo"
                    },
                    "direccion": {
                        "type": "string",
                        "example": "Parque Nacional Cotopaxi"
                    },
                    "latitud": {
                        "type": "number",
                        "format": "float",
                        "example": "-0.6836"
                    },
                    "longitud": {
                        "type": "number",
                        "format": "float",
                        "example": "-78.4372"
                    },
                    "horario_visitas": {
                        "type": "string",
                        "example": "08:00 - 17:00"
                    },
                    "precio_entrada": {
                        "type": "number",
                        "format": "float",
                        "example": 10.5
                    },
                    "estado": {
                        "type": "boolean",
                        "example": true
                    },
                    "categoria_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "provincia_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}